requirements.txt
asgiref==3.5.2
black==22.8.0
click==8.1.3
Django==4.0.5
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
gunicorn==20.1.0
importlib-metadata==4.12.0
Markdown==3.4.1
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
psycopg2-binary==2.9.3
pytz==2022.2.1
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.3.0
whitenoise==6.2.0
zipp==3.8.1
